requirements.txt
streamlit
pillow
pandas
torch>=2.0
transformers
faiss-cpu